feat(11.1): migrate solana-sdk 1.18 → 4.x
- Bump solana-sdk to 4, solana-client to 4.0.0-beta.7 (user-accepted beta risk per OQ-1)
- Bump borsh 0.10 → 1 to satisfy solana-address::Address BorshDeserialize/Serialize (D-11 reactive fix)
- No changes needed in 4 of 5 planned src files — solana_sdk::pubkey::{Pubkey, pubkey!} and
  solana_client::rpc_client::RpcClient re-exports remain stable in 4.x (confirming RESEARCH H-confidence).
- Minor clippy fixes for new rustc 1.94 lints triggered by toolchain bump:
  - src/rpc.rs: vec![] → [] in test helper (clippy::useless_vec)
  - src/math/fees.rs: parenthesize (a_lo * b_lo) >> 64 (clippy::precedence)
  - src/backtest/mod.rs: allow(clippy::items_after_test_module) for pre-existing layout
  - src/backtest/db_replay.rs: drop unnecessary u128 → f64 casts (clippy::unnecessary_cast)
- All 350 unit tests green; cargo clippy --all-targets -- -D warnings clean.
- audit.toml NOT touched — cleanup scheduled in Plan 04.

Bisect-safe: cargo build --all-targets && cargo test --all && cargo clippy --all-targets -- -D warnings green on this SHA (D-12).

+++ src/backtest/db_replay.rs
@@ -349,7 +349,7 @@ mod tests {
         let input = default_input(); // position_liquidity = 100
         let result = run_db_backtest(input, &ticks).unwrap();
 
-        let expected_fees = (100_u128 as f64 / 18_446_744_073_709_551_616.0_f64) * 0.1;
+        let expected_fees = (100_f64 / 18_446_744_073_709_551_616.0_f64) * 0.1;
         assert!(
             (result.total_fees_usd - expected_fees).abs() < 1e-20,
             "expected fees ~{:.2e}, got {:.2e}",
@@ -370,7 +370,7 @@ mod tests {
         let result = run_db_backtest(input, &ticks).unwrap();
 
         // delta = wrapping_sub(5, u128::MAX) = 6
-        let expected_fees = (6_u128 as f64 / 18_446_744_073_709_551_616.0_f64) * 0.1;
+        let expected_fees = (6_f64 / 18_446_744_073_709_551_616.0_f64) * 0.1;
         assert!(
             (result.total_fees_usd - expected_fees).abs() < 1e-30,
             "expected wrapping fees ~{:.2e}, got {:.2e}",
